start of oc past max

+++ src/main/java/gregtech/api/capability/impl/MultiblockRecipeLogic.java
@@ -75,7 +75,7 @@ public class MultiblockRecipeLogic extends AbstractRecipeLogic {
     }
 
     @Override
-    protected boolean drawEnergy(int recipeEUt) {
+    protected boolean drawEnergy(long recipeEUt) {
         long resultEnergy = getEnergyStored() - recipeEUt;
         if (resultEnergy >= 0L && resultEnergy <= getEnergyCapacity()) {
             getEnergyContainer().changeEnergy(-recipeEUt);
