+++ src/main/java/gregtech/api/capability/impl/RecipeLogicEnergy.java
@@ -29,7 +29,7 @@ public class RecipeLogicEnergy extends AbstractRecipeLogic {
     }
 
     @Override
-    protected boolean drawEnergy(int recipeEUt) {
+    protected boolean drawEnergy(long recipeEUt) {
         long resultEnergy = getEnergyStored() - recipeEUt;
         if (resultEnergy >= 0L && resultEnergy <= getEnergyCapacity()) {
             energyContainer.get().changeEnergy(-recipeEUt);
