+++ src/main/java/gregtech/api/capability/impl/RecipeLogicSteam.java
@@ -163,10 +163,10 @@ public class RecipeLogicSteam extends AbstractRecipeLogic {
     }
 
     @Override
-    protected int[] calculateOverclock(int EUt, long voltage, int duration) {
+    protected long[] calculateOverclock(long EUt, long voltage, int duration) {
         if (!isHighPressure) {
             //disallow overclocking for low pressure bronze machines
-            return new int[]{EUt, duration};
+            return new long[]{EUt, duration};
         }
         return super.calculateOverclock(EUt, voltage, duration);
     }
@@ -182,7 +182,7 @@ public class RecipeLogicSteam extends AbstractRecipeLogic {
     }
 
     @Override
-    protected boolean drawEnergy(int recipeEUt) {
+    protected boolean drawEnergy(long recipeEUt) {
         int resultDraw = (int) Math.ceil(recipeEUt / conversionRate);
         return resultDraw >= 0 && steamFluidTank.getFluidAmount() >= resultDraw &&
             steamFluidTank.drain(resultDraw, true) != null;
